fixed bugs
We have implemented functionality that is not shown in the Demo video. Primarily, we updated the bug where the app would match a user with oneself. Furthermore, we fixed the bug where it showed the same zip code as being far away from each other(which is in the video). Lastly, we fixed a bug causing the site to crash when no matches were found, and we expanded our database of medicines.